+++ Config.h
@@ -45,6 +45,7 @@
 
 //=================================================================
 #include <ESP8266WiFi.h>
+#include "WiFiManager.h"
 #include <ESP8266HTTPClient.h>
 #include "ESP8266WebServer.h"
 #include "ESP8266httpUpdate.h"
@@ -197,6 +198,9 @@ char moduleName[20] = "Test";
 WiFiClient espClient;
 HTTPClient http;
 PubSubClient client(espClient);
+
+// Create an instance of the WiFiManager class
+WiFiManager wifiManager(wifi_ssid, wifi_password, wifi_StaticIP, wifi_gateway, wifi_subnet, wifi_DNS, wifi_hostname, wifi_static, softAP_ssid, softAP_pass);
 
 #if ( EXCLUDED_CODE == 9 )  //===============================================
 // TEST WITH SSL