+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **000.05.105** - in development
 - added CPU Frequency(MHz) in deviceinfo.json
 - cleanUP Main Loop - moved logic to each owner
+- WiFiManager moved to Class and new file WiFiManager.h
 
 **000.05.104**  
 - wifiScanJSON() rewritten, using asyn method WiFi.scanNetworksAsync(wifiScanResult)